refactor(usa-apply): align apply module structure with core
