+++ legify-web-usa/projects/apply/src/lib/models/application/index.ts
@@ -1,0 +1,1 @@
+export * from './legify-usa-application';
