+++ legify-web-usa/projects/apply/src/lib/models/index.ts
@@ -1,0 +1,1 @@
+export * from './application';
