+++ legify-web-usa/projects/apply/src/lib/modules/usa-apply-basic-info/usa-apply-basic-info-routing.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+const routes: Routes = [];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class UsaApplyBasicInfoRoutingModule { }
