+++ legify-web-usa/projects/apply/src/lib/modules/usa-apply-basic-info/usa-apply-basic-info.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { UsaApplyBasicInfoRoutingModule } from './usa-apply-basic-info-routing.module';
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    UsaApplyBasicInfoRoutingModule
+  ]
+})
+export class UsaApplyBasicInfoModule { }
