+++ legify-web-usa/projects/apply/src/lib/modules/usa-apply-documents/services/index.ts
@@ -1,0 +1,1 @@
+export * from './usa-apply-documents/usa-apply-documents.service';
