+++ legify-web-usa/projects/apply/src/lib/modules/usa-apply-documents/services/usa-apply-documents/usa-apply-documents.service.ts
@@ -13,7 +13,7 @@ import {
 } from '@legify/web-apply';
 import { Observable } from 'rxjs';
 import { map, withLatestFrom } from 'rxjs/operators';
-import { UsaLegifyApplication } from '../../../models/application/legify-usa-application';
+import { UsaLegifyApplication } from '../../../../models';
 import { USA_SUPPORTING_DOC_TYPE_GROUP } from '../../constants';
 
 @Injectable()
