--- legify-web-usa/projects/apply/src/lib/usa-apply-documents/index.ts
@@ -1,1 +1,0 @@
-export * from './public-api';
