+++ legify-web-usa/projects/apply/src/public-api.ts
@@ -2,7 +2,8 @@
  * Public API Surface of apply
  */
 
+export * from './lib/models';
+export * from './lib/modules';
 export * from './lib/services';
 export * from './lib/usa-apply.module';
 export * from './lib/usa-apply.component';
-export * from './lib/usa-apply-documents';
